Correction bug seulement Linux déplacement image dans modèle

+++ noethys/Dlg/DLG_Noedoc.py
@@ -1005,7 +1005,7 @@ def AjouterPolygone( points=[], nom=_(u"Polygone"), champ=None,
 
 def AjouterImage(bmp, xy, hauteur=None, nom=_(u"Image"), champ=None, typeImage="fichier", IDobjet=None, InForeground=True):
     """ Création d'une image """
-    objet = MovingScaledBitmap(bmp, xy, Height=hauteur, Position="bl", InForeground=InForeground)
+    objet = MovingScaledBitmap(bmp, xy.astype(float), Height=hauteur, Position="bl", InForeground=InForeground)
     objet.nom = nom
     objet.champ = champ
     objet.categorie = "image"
@@ -3095,7 +3095,7 @@ class Panel_canvas(wx.Panel):
             self.Selection(objet) 
         else:
             if self.dictSelection["objet"] != objet :
-                self.Selection(objet) 
+                self.Selection(objet)
 
     def OnClicDroitObjet(self, objet):
         """ Clic droit sur objet """
@@ -4747,7 +4747,7 @@ def ImportationObjets(IDmodele=None, InForeground=True):
                     InForeground=InForeground,
                     )
         
-        # Ligne de texte
+        # Bloc de texte
         if objet["categorie"] == "bloc_texte" :
             objetCanvas = AjouterBlocTexte(
                     objet["texte"],
